Delegate grid management to Lattice module for all blocks
- Major changes to the world / simulator files
- Global refactoring of grid management for all blocks
- A few bugs still need fixing, working on it
- Further refactoring of this files to come

+++ applicationsSrc/robot01/robot01BlockCode.h
@@ -51,7 +51,7 @@ typedef boost::shared_ptr<RunTrainMessage> RunTrainMessage_ptr;
 */
 
 class Robot01BlockCode : public RobotBlocks::RobotBlocksBlockCode {
-	int gridSize[3];
+	short gridSize[3];
 	RobotBlocks::presence *targetGrid;
 	P2PNetworkInterface *block2Answer;
 	int nbreOfWaitedAnswers,blockToUnlock;
@@ -92,9 +92,9 @@ public:
 
 class MapMessage : public Message {
 public :
-	int gridSize[3];
+	short gridSize[3];
 	RobotBlocks::presence *targetGrid;
-	MapMessage(int*,RobotBlocks::presence*);
+	MapMessage(short*,RobotBlocks::presence*);
 	~MapMessage();
 };
